update code on may 25

+++ src/app/device/routers/routers.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import {Router,ActivatedRoute} from '@angular/router';
 import { BackendconnectionService } from '../../backendconnection.service';
+import Swal from 'sweetalert2/dist/sweetalert2.js'; 
 
 @Component({
   selector: 'app-routers',
@@ -17,6 +18,12 @@ export class RoutersComponent implements OnInit {
   id:any;
 
   ngOnInit(): void {
+   
+    this.getdata();
+
+  }
+
+  getdata(){
     this.route.paramMap.subscribe(params => {
       this.id = params.get("id");
       })
@@ -28,7 +35,6 @@ export class RoutersComponent implements OnInit {
         
   
       });
-
   }
 
   addrouter(x = this.id){
@@ -40,11 +46,54 @@ export class RoutersComponent implements OnInit {
   }
 
 
+  // delete(x,y){
+  //   this.backend.deleterouter(x,y)
+  //   .subscribe((data)=> { 
+  //       console.log(data);
+  //   });
+  // }
+
   delete(x,y){
-    this.backend.deleterouter(x,y)
+
+    //  alert(x)
+    // if (confirm('Are you sure to delete this record ?') == true) {
+      
+  
+   
+      Swal.fire({
+        title: 'Are you sure?',
+        text: 'You will not be able to recover this Router!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonText: 'Yes, delete it!',
+        cancelButtonText: 'No, keep it'
+      }).then((result) => {
+        if (result.value) {
+  
+           this.backend.deleterouter(x,y)
     .subscribe((data)=> { 
+  
         console.log(data);
+      this.getdata();
+      Swal.fire(
+        'Deleted!',
+        'Router has been deleted.',
+        'success'
+      )
     });
+          
+        // For more information about handling dismissals please visit
+        // https://sweetalert2.github.io/#handling-dismissals
+        } else if (result.dismiss === Swal.DismissReason.cancel) {
+          Swal.fire(
+            'Cancelled',
+            'Router is safe :)',
+            'error'
+          )
+        }
+      }) 
+   
+  // }
   }
 
   edit(x,y){
